library_system/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
mysqlclient==2.2.7
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
